Some method about booking schedule.

+++ backend/routes/api.php
@@ -36,9 +36,9 @@ Route::get('booking/index','Api\BookingController@index')->middleware('auth:api'
 Route::get('/booking/show/{id}','Api\BookingController@show')->middleware('auth:api');
 Route::post('/booking/store', 'Api\BookingController@store')->middleware('auth:api');
 Route::patch('/booking/update/{id}','Api\BookingController@update')->middleware('auth:api');
-Route::delete('booking/destroy/{id}','Api\BookingController@destroy');
-
-
+Route::delete('booking/destroy/{id}','Api\BookingController@destroy')->middleware('auth:api');
+Route::get('/booking/cancel-booking/{id}','Api\BookingController@cancel_booking')->middleware('auth:api');
+Route::get('/booking/show-cancel-booking','Api\BookingController@show_cancel_booking')->middleware('auth:api');
